# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tagg@d API",
    "version": "1.0.0",
    "description": "API documentation for the Tagg@d backend — a crypto payment platform for Africa."
  },
  "servers": [
    {
      "url": "http://localhost:5002/api/v2",
      "description": "Current version (v2)"
    },
    {
      "url": "http://localhost:5002/api/v1",
      "description": "Deprecated version (v1)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Development Server (unversioned root)"
    },
    {
      "url": "https://taggedpay.xyz/api/v2",
      "description": "Production (v2)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/api/api-keys": {
      "post": {
        "summary": "Create a new API key",
        "description": "Generate an API key with specified scopes for third-party access. The key is returned only once — store it securely.",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 100,
                    "example": "Partner Integration Key"
                  },
                  "scopes": {
                    "type": "string",
                    "default": "read,write",
                    "description": "Comma-separated scopes (transactions:read, transactions:write, payments:send, webhooks:read, webhooks:write)",
                    "example": "transactions:read,webhooks:read"
                  },
                  "ipWhitelist": {
                    "type": "string",
                    "description": "Comma-separated IP addresses or CIDR ranges",
                    "example": "203.0.113.0/24,192.168.1.1"
                  },
                  "expiresIn": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 365,
                    "description": "Number of days until the key expires",
                    "example": 90
                  },
                  "rotationIntervalDays": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 365,
                    "description": "Automatic rotation interval in days",
                    "example": 30
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "name": {
                          "type": "string",
                          "example": "Partner Integration Key"
                        },
                        "key": {
                          "type": "string",
                          "example": "tagged_live_sk_a1b2c3d4e5f6..."
                        },
                        "scopes": {
                          "type": "string",
                          "example": "transactions:read,webhooks:read"
                        },
                        "expires_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      },
      "get": {
        "summary": "List all API keys for the authenticated user",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "scopes": {
                            "type": "string"
                          },
                          "is_active": {
                            "type": "boolean"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "summary": "Login user and get JWT token",
        "description": "Authenticate with email/password. If 2FA is enabled, include the `twoFactorToken`. Returns a JWT bearer token.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.lagos@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "StrongP@ssw0rd!"
                  },
                  "twoFactorToken": {
                    "type": "string",
                    "example": "123456",
                    "description": "6-digit OTP from authenticator app (required if 2FA is enabled)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "example": "user"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          },
          "401": {
            "description": "Unauthorized or 2FA token required"
          }
        }
      }
    },
...
  },
...
}